Adjust calibration process to use set points (#8)
The existing calibration process seems finicky across different bed
types. Feedback has indicated that the value ends up a little high,
resulting in false negatives. On the other side, some beds display
a rebound effect, where the sensor goes to a very low value when someone
gets out of bed, but then slowly increases over the next couple hours.
This rebound effect requires that the threshold percentage be set pretty
high. To remedy this, the calibration process now allows you to set
unoccupied and occupied values and calculates the midpoint between the
two. For beds that display 0 when unoccupied, it allows the calculated
value to be much lower. For beds that rebound, it allows the unoccupied
state to be set at the fully rebounded state instead of the value it
reports when someone has just gotten out of bed.

* Adjust calibration process to use set points
* Default trigger_percentile to 0.50
* Move pressure status to numbers